Tests implementiert

+++ src/main/java/item/service/PersonalItemService.java
@@ -16,8 +16,20 @@ public class PersonalItemService {
     }
 
     public PersonalItem create(PersonalItem item) {
-        // TODO item data check
-        return repository.add(item);
+        if(checkPersonalItemData(item)) {
+            return repository.add(item);
+        }
+        throw new ValidationError("");
+    }
+
+    private boolean checkPersonalItemData(PersonalItem item) {
+        if(item.getNachname() == null || item.getNachname().isEmpty()) {
+            throw new ValidationError("Nachname darf nicht leer sein.");
+        }
+        if(item.getVorname() == null || item.getVorname().isEmpty()) {
+            throw new ValidationError("Vorname darf nicht leer sein.");
+        }
+        return true;
     }
 
     public PersonalItem getById(Long id) {
@@ -25,12 +37,18 @@ public class PersonalItemService {
     }
 
     public Boolean update(PersonalItem item) {
-        // TODO item data check
-        repository.update(item);
-        return true;
+        if(checkPersonalItemData(item)) {
+            repository.update(item);
+            return true;
+        }
+        return false;
     }
 
     public boolean delete(long id) {
-        return repository.delete(id);
+        if(id>0) {
+            return repository.delete(id);
+        } else {
+            throw new ValidationError("ID muss grösser als 0 sein.");
+        }
     }
 }
